Move ControllerPollProperties -> org.eclipse.hawkbit.tenancy.configuration
_release_notes_
